--- src/components/OverviewComponent/OverviewComponent.test.tsx
@@ -1,8 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import OverviewComponent from './OverviewComponent';
-
-test('renders learn react link', () => {
-  render(<OverviewComponent />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
