+++ src/components/atoms/card/card.tsx
@@ -1,6 +1,6 @@
 import { Card, CardActionArea, CardContent, Typography } from "@mui/material";
 import { Chore } from "../../../models/Chore";
-import './Card.css';
+import './card.css';
 
 function CardListItem({ id, title, description }: Chore) {
   return (
