--- src/models/CardListProps.tsx
@@ -1,5 +1,0 @@
-import { Chore } from "./Chore";
-
-export interface CardListProps {
-    chores: Chore[];
-}
